Add files via upload
Decomposed some repeated code into actual_input_length in reviewer.h.

After setting to read-only in a CHERI environment, set the original reviewer to the read-only reviewer that we created (so that the subsequent buffer overflow attacks will be on the read-only reviewer). 
OPEN ISSUE: this may be misleading, as it doesn't really make a difference in a CHERI environment whether or not the struct is read-only - it will crash either way.

Fixed bug in which the conditional CHERI environment block before setting the reviewer struct to read-only, was unreachable.